third_party/WebKit/LayoutTests/http/tests/websocket/handshake-fail-by-mismatch-protocol-header-expected.txt

CONSOLE ERROR: line 22: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: Response must not include 'Sec-WebSocket-Protocol' header if not present in request: MismatchProtocol
CONSOLE ERROR: line 24: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: 'Sec-WebSocket-Protocol' header value 'MismatchProtocol' in response does not match any of sent values
CONSOLE ERROR: line 24: WebSocket connection to 'ws://localhost:8880/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: 'Sec-WebSocket-Protocol' header value 'MismatchProtocol' in response does not match any of sent values
Test whether WebSocket handshake fails if the server sends mismatched Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS closeEvent.wasClean is false
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
